[HtmlSanitizer][FrameworkBundle] Add HtmlSanitizerBundle
(cherry picked from commit d8e04bc92e841359abbbabba0f198d6e42ad05ca)
