+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/RemoveUnusedFormHtmlSanitizerPass.php
@@ -1,0 +1,30 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+/**
+ * Removes the form type extension that sanitizes HTML when no sanitizer is registered.
+ *
+ * @internal
+ */
+class RemoveUnusedFormHtmlSanitizerPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        if (!$container->has('html_sanitizer')) {
+            $container->removeDefinition('form.type_extension.form.html_sanitizer');
+        }
+    }
+}
