+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -27,7 +27,6 @@ use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Exception\LogicException;
 use Symfony\Component\Form\Form;
-use Symfony\Component\HtmlSanitizer\HtmlSanitizerInterface;
 use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\IpUtils;
@@ -188,7 +187,7 @@ class Configuration implements ConfigurationInterface
         $this->addNotifierSection($rootNode, $enableIfStandalone);
         $this->addRateLimiterSection($rootNode, $enableIfStandalone);
         $this->addUidSection($rootNode, $enableIfStandalone);
-        $this->addHtmlSanitizerSection($rootNode, $enableIfStandalone);
+        $this->addHtmlSanitizerSection($rootNode);
         $this->addWebhookSection($rootNode, $enableIfStandalone);
         $this->addRemoteEventSection($rootNode);
         $this->addJsonStreamerSection($rootNode, $enableIfStandalone);
@@ -2599,140 +2598,14 @@ class Configuration implements ConfigurationInterface
         ;
     }
 
-    /**
-     * @param-immediately-invoked-callable $enableIfStandalone
-     */
-    private function addHtmlSanitizerSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
+    private function addHtmlSanitizerSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
-                ->arrayNode('html_sanitizer')
-                    ->info('HtmlSanitizer configuration')
-                    ->{$enableIfStandalone('symfony/html-sanitizer', HtmlSanitizerInterface::class)}()
-                    ->children()
-                        ->arrayNode('sanitizers', 'sanitizer')
-                            ->useAttributeAsKey('name')
-                            ->arrayPrototype()
-                                ->children()
-                                    ->enumNode('default_action')
-                                        ->info('Defines how the sanitizer must behave by default.')
-                                        ->values(['drop', 'block', 'allow'])
-                                    ->end()
-                                    ->booleanNode('allow_safe_elements')
-                                        ->info('Allows "safe" elements and attributes.')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->booleanNode('allow_static_elements')
-                                        ->info('Allows all static elements and attributes from the W3C Sanitizer API standard.')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->arrayNode('allow_elements', 'allow_element')
-                                        ->info('Configures the elements that the sanitizer should retain from the input. The element name is the key, the value is either a list of allowed attributes for this element or "*" to allow the default set of attributes (https://wicg.github.io/sanitizer-api/#default-configuration).')
-                                        ->example(['i' => '*', 'a' => ['title'], 'span' => 'class'])
-                                        ->normalizeKeys(false)
-                                        ->useAttributeAsKey('name')
-                                        ->variablePrototype()
-                                            ->beforeNormalization()
-                                                ->ifArray()->then(static fn ($n) => $n['attribute'] ?? $n)
-                                            ->end()
-                                            ->validate()
-                                                ->ifTrue(static fn ($n): bool => !\is_string($n) && !\is_array($n))
-                                                ->thenInvalid('The value must be either a string or an array of strings.')
-                                            ->end()
-                                        ->end()
-                                    ->end()
-                                    ->arrayNode('block_elements', 'block_element')
-                                        ->info('Configures elements as blocked. Blocked elements are elements the sanitizer should remove from the input, but retain their children.')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->arrayNode('drop_elements', 'drop_element')
-                                        ->info('Configures elements as dropped. Dropped elements are elements the sanitizer should remove from the input, including their children.')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->arrayNode('allow_attributes', 'allow_attribute')
-                                        ->info('Configures attributes as allowed. Allowed attributes are attributes the sanitizer should retain from the input.')
-                                        ->normalizeKeys(false)
-                                        ->useAttributeAsKey('name')
-                                        ->variablePrototype()
-                                            ->beforeNormalization()
-                                                ->ifArray()->then(static fn ($n) => $n['element'] ?? $n)
-                                            ->end()
-                                        ->end()
-                                    ->end()
-                                    ->arrayNode('drop_attributes', 'drop_attribute')
-                                        ->info('Configures attributes as dropped. Dropped attributes are attributes the sanitizer should remove from the input.')
-                                        ->normalizeKeys(false)
-                                        ->useAttributeAsKey('name')
-                                        ->variablePrototype()
-                                            ->beforeNormalization()
-                                                ->ifArray()->then(static fn ($n) => $n['element'] ?? $n)
-                                            ->end()
-                                        ->end()
-                                    ->end()
-                                    ->arrayNode('force_attributes', 'force_attribute')
-                                        ->info('Forcefully set the values of certain attributes on certain elements.')
-                                        ->normalizeKeys(false)
-                                        ->useAttributeAsKey('name')
-                                        ->arrayPrototype()
-                                            ->normalizeKeys(false)
-                                            ->useAttributeAsKey('name')
-                                            ->stringPrototype()->end()
-                                        ->end()
-                                    ->end()
-                                    ->booleanNode('force_https_urls')
-                                        ->info('Transforms URLs using the HTTP scheme to use the HTTPS scheme instead.')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->arrayNode('allowed_link_schemes', 'allowed_link_scheme')
-                                        ->info('Allows only a given list of schemes to be used in links href attributes.')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->arrayNode('allowed_link_hosts', 'allowed_link_host')
-                                        ->info('Allows only a given list of hosts to be used in links href attributes.')
-                                        ->defaultNull()
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->booleanNode('allow_relative_links')
-                                        ->info('Allows relative URLs to be used in links href attributes.')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->arrayNode('allowed_media_schemes', 'allowed_media_scheme')
-                                        ->info('Allows only a given list of schemes to be used in media source attributes (img, audio, video, ...).')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->arrayNode('allowed_media_hosts', 'allowed_media_host')
-                                        ->info('Allows only a given list of hosts to be used in media source attributes (img, audio, video, ...).')
-                                        ->defaultNull()
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->booleanNode('allow_relative_medias')
-                                        ->info('Allows relative URLs to be used in media source attributes (img, audio, video, ...).')
-                                        ->defaultFalse()
-                                    ->end()
-                                    ->arrayNode('with_attribute_sanitizers', 'with_attribute_sanitizer')
-                                        ->info('Registers custom attribute sanitizers.')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->arrayNode('without_attribute_sanitizers', 'without_attribute_sanitizer')
-                                        ->info('Unregisters custom attribute sanitizers.')
-                                        ->acceptAndWrap(['string'])
-                                        ->stringPrototype()->end()
-                                    ->end()
-                                    ->integerNode('max_input_length')
-                                        ->info('The maximum length allowed for the sanitized input.')
-                                        ->defaultValue(0)
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
+                ->variableNode('html_sanitizer')
+                    ->aliasOf('html_sanitizer')
+                    ->treatFalseLike(['enabled' => false])
+                    ->treatTrueLike(['enabled' => true])
                 ->end()
             ->end()
         ;
