+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/full.php
@@ -81,7 +81,4 @@ $container->loadFromExtension('framework', [
             'pdf' => 'application/pdf',
         ],
     ],
-    'html_sanitizer' => [
-        'enabled' => true,
-    ],
 ]);
