--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/html_sanitizer_default_allowed_link_and_media_hosts.php
@@ -1,9 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'html_sanitizer' => [
-        'sanitizers' => [
-            'custom_default' => null,
-        ],
-    ],
-]);
