--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/html_sanitizer_default_config.php
@@ -1,4 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'html_sanitizer' => null]);
