+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/legacy_html_sanitizer.php
@@ -1,0 +1,14 @@
+<?php
+
+return function (Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator $container) {
+    $container->services()->alias('test_html_sanitizer', 'html_sanitizer.sanitizer.custom')->public();
+    $container->extension('framework', [
+        'html_sanitizer' => [
+            'sanitizers' => [
+                'custom' => [
+                    'default_action' => 'block',
+                ],
+            ],
+        ],
+    ]);
+};
