--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/html_sanitizer_default_allowed_link_and_media_hosts.yml
@@ -1,4 +1,0 @@
-framework:
-    html_sanitizer:
-        sanitizers:
-            custom_default: ~
