--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/html_sanitizer_default_config.yml
@@ -1,2 +1,0 @@
-framework:
-    html_sanitizer: ~
