+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -52,7 +52,7 @@
         "symfony/dotenv": "^7.4|^8.0",
         "symfony/expression-language": "^7.4|^8.0",
         "symfony/form": "^7.4|^8.0",
-        "symfony/html-sanitizer": "^7.4|^8.0",
+        "symfony/html-sanitizer": "^8.2",
         "symfony/http-client": "^8.2",
         "symfony/json-streamer": "^7.4|^8.0",
         "symfony/lock": "^7.4|^8.0",
@@ -89,6 +89,7 @@
         "phpdocumentor/type-resolver": "<1.5.1",
         "symfony/console": "<8.1.2",
         "symfony/form": "<7.4",
+        "symfony/html-sanitizer": "<8.2",
         "symfony/http-client": "<8.2",
         "symfony/json-path": "<8.2",
         "symfony/json-streamer": "<7.4",
