+++ src/Symfony/Component/HtmlSanitizer/Resources/config/html_sanitizer.php
