+++ src/Symfony/Component/HtmlSanitizer/Tests/DependencyInjection/Fixtures/html_sanitizer_default_allowed_link_and_media_hosts.php
@@ -1,0 +1,7 @@
+<?php
+
+$container->loadFromExtension('html_sanitizer', [
+    'sanitizers' => [
+        'custom_default' => null,
+    ],
+]);
