+++ src/Symfony/Component/HtmlSanitizer/Tests/DependencyInjection/Fixtures/html_sanitizer_default_config.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('html_sanitizer', null);
